Trying to remove comdat from non-Windows part of build
There are a number of modifers like linkonce and weak that *might* be enough to merge identical LLVM constants together at link time. If so, it *might* solve the problem with building on MacOs, which doesn't support comdat. This commit takes out comdat, but it doesn't work, complaining of "fatal error LNK1227: conflicting weak extern definition."