Changed instantiated generics to get saved explicitly in the type
Although generic types are held in the type, they are added to the type in the TAC generation process so that only TACClass objects that require a generic type trigger a generic type storage.

The exception to this is any type that's being read from a .meta file, since the .meta file has a list of fully-instantiated generics at the end.

The goal of this approach is to have shorter lists of generic types in both .meta files and instantiated in the final executable. A second benefit is less time spent cycling through the set of usedTypes.